requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.7
django-extensions==2.2.6
django-filter==2.2.0
django-guardian==2.1.0
django-heroku==0.3.1
django-picklefield==2.1.1
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.3.0
itypes==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==8.0.2
mysql-connector-python==8.0.13
numpy==1.17.2
packaging==20.0
pipenv==2018.11.26
pluggy==0.13.1
protobuf==3.11.2
psycopg2==2.8.4
py==1.8.1
Pygments==2.5.2
pyparsing==2.4.2
pytest==5.3.2
python-dateutil==2.8.0
python-decouple==3.3
pytz==2019.3
requests==2.22.0
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==16.7.9
virtualenv-clone==0.5.3
wcwidth==0.1.8
whitenoise==5.0.1
zipp==0.6.0